Fixed student mutation - is working
Mutation works but gives a non-nullable error. I figured the mutation
was written correctly but it kept complaining about how I was giving the
input variables in the gql playground. I changed how the resolve looks
to copy the same format of the ones in schema.ts file. I console logged
inside the resolve to see what the args object contained which was all
the right fields. The main issue was the enums because I was defining
them as ints instead of strings. So graphql didn't know how to interpret
the ints and needed to give strings. Changed the migration as well to
ignore email field.