Fixed slides query and mutation on a lesson
Currently all slides can be shown for a lesson upon clicking on its edit
button. Also, slides sidebar gets updated when adding a new slide.
However, updating a slide to a multiple choice question block or a text
block brings up errors. Also, orphan data is left when deleting a
lesson. Database still holds all slides and blocks made for that lesson,
but are no longer attached to a lesson. Needs to be deleted.