Fixed slides query and mutation on slides page
Had to change the added layer of querying on a lesson and slide for the
mutations. For example, anywhere that said "mutation mut( ) { lesson
{...}...}" had to be changed to just "mutation mut ( ) { ... }".
Basically to just return the fields of the object directly instead of
returning the object - it was an unnecessary added layer of abstraction
that David Castaneda pointed out to us. Also had to change how the
MCQuestBlock and TextBlock were returning its fields. Had to use prisma
in order to return its fields.