Fixed refreshing of lessonplan editing
Had to destructure the courses field of a course object. On prisma, this
courses field is a list of CourseToLesson objects that hold the
connections between a course and lesson. But it had an extra layer of
wrapping with courses: { lesson: {..}} but we want just lessons: {...}.
The course gql object has its lessons field resolve updated to do just
this.