Fix delete course button to delete course onsubmit
Updated the delete course mutation to delete all lessons and students
associated with the course before deleting the course. Updated its
return to include all the needed associated info such as lessons and
students that were in this course. Upon delete course mutation success,
the history is updated and directs user to the courses page.