Fix undefined instructor when gradebook is clicked
Had to change where the query is being called. seems like the query
lagged behind sometimes when trying to get the id when the sidebar first
comes into view. seems to work better
with a query called everytime when gradebook or grades is clicked on and
gets the viewer id that way. This might not be the best solution since
it's doing a query every single time a tab is clicked on. Would be the
best to do a one time query when the user logs in and save that viewer
id globally.